Utility methods from CloneGitRepositoryTask moved to CloneGitRepositoryTaskFactory
