+++ subprojects/shipkit/src/test/groovy/org/shipkit/internal/gradle/git/tasks/CloneGitRepositoryTaskTest.groovy
@@ -30,50 +30,4 @@ class CloneGitRepositoryTaskTest extends PluginSpecification {
         expect:
         task.getCloneCommand() == ["git", "clone", "--depth", "50", "url", tempFolder.getAbsolutePath()]
     }
-
-    def "returns correct consumerRepoCloneDir"() {
-        def buildDir = tmp.newFolder("buildTest")
-        project.setBuildDir(buildDir)
-
-        when:
-        def result = CloneGitRepositoryTask.getConsumerRepoCloneDir(project, "mockito/shipkit-example")
-
-        then:
-        result.absolutePath == buildDir.absolutePath + "/downstream/mockitoShipkitExample"
-    }
-
-    def "configures cloneRepoName task"() {
-        given:
-        def buildDir = tmp.newFolder("buildTest")
-        project.setBuildDir(buildDir)
-
-        when:
-        def cloneTask = CloneGitRepositoryTask.createCloneTask(project, "gitHubUrl", "mockito/mockito")
-
-        then:
-        cloneTask.repositoryUrl == "gitHubUrl/mockito/mockito"
-        cloneTask.targetDir.absolutePath == buildDir.absolutePath + "/downstream/mockitoMockito"
-        cloneTask.name == "cloneMockitoMockito"
-    }
-
-    def "creates only one instance of cloneRepoName"() {
-        when:
-        def one = CloneGitRepositoryTask.createCloneTask(project, "gitHubUrl", "mockito/mockito")
-        def another = CloneGitRepositoryTask.createCloneTask(project, "gitHubUrl", "mockito/mockito")
-
-        then:
-        one == another
-    }
-
-    def "creates cloneRepoName task in root project"() {
-        given:
-        def child = new ProjectBuilder().withParent(project).withName("child").build()
-
-        when:
-        CloneGitRepositoryTask.createCloneTask(child, "gitHubUrl", "mockito/mockito")
-
-        then:
-        child.tasks.findByName("cloneMockitoMockito") == null
-        project.tasks.cloneMockitoMockito
-    }
 }
